fix(eslint-plugin): recognize extension-registered annotation tags (#350)
tag-recognition/no-unknown-tags and tag-recognition/tsdoc-comment-syntax
now iterate extension.annotations alongside constraintTags and
metadataSlots, so tags declared via defineAnnotation() (e.g.
@primaryField) are treated as known. Closes #350.
